rename analyze_models => models_reports
